@@ -20,8 +20,8 @@
 		"@types/node": "^20.10.6",
 		"@types/ws": "^8.5.10",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"svelte-check": "^3.6.0"
+		"prettier-plugin-svelte": "^3.2.6",
+		"svelte-check": "^4.0.0"
 	},
 	"type": "module",
 	"dependencies": {
@@ -30,8 +30,8 @@
 		"@napi-rs/canvas": "^0.1.44",
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-node": "^2.0.2",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"axios": "^1.6.4",
 		"cloud-url-resolver": "^1.3.0",
 		"discord.js": "^14.14.1",
@@ -39,13 +39,13 @@
 		"ms": "^2.1.3",
 		"onnxruntime-node": "^1.16.3",
 		"sharp": "^0.33.1",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0",
 		"svelte-fa": "^4.0.2",
 		"svelte-french-toast": "^1.2.0",
 		"systeminformation": "^5.21.24",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4",
 		"vitest": "^1.0.0",
 		"ws": "^8.16.0",
 		"autoprefixer": "^10.4.16",
