@@ -1,9 +1,14 @@
 <script lang="ts">
 	import Fa from 'svelte-fa';
 	import { faCaretDown, faCaretLeft, type IconDefinition } from '@fortawesome/free-solid-svg-icons';
-	export let title: string;
-	export let icon: IconDefinition;
-	let minimized = false;
+	interface Props {
+		title: string;
+		icon: IconDefinition;
+		children?: import('svelte').Snippet;
+	}
+
+	let { title, icon, children }: Props = $props();
+	let minimized = $state(false);
 
 	function toggleMinimize() {
 		minimized = !minimized;
@@ -13,7 +18,7 @@
 <div class="window">
 	<div class="title-bar">
 		<h2 class="title"><span class="title-icon"><Fa {icon} /></span>{title}</h2>
-		<button on:click={toggleMinimize} class="title-button">
+		<button onclick={toggleMinimize} class="title-button">
 			{#if minimized}
 				<Fa icon={faCaretLeft} />
 			{:else}
@@ -23,7 +28,7 @@
 	</div>
 	<div class="content">
 		{#if !minimized}
-			<slot />
+			{@render children?.()}
 		{/if}
 	</div>
 </div>
